feat: extracted top and sidebar into components
